Extensions: Expose inactive tab color
Currently, the inactive tab color is the same as the regular text
color. For most themers, this isn't an issue but some would like to
highlight the currently selected tab with an accent and that accent
carries throughout the rest of the text colors.

Before: http://i.imgur.com/k3t9jyO.png
After: http://i.imgur.com/OFDJcNe.png

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Change-Id: Iceede426936bbf6effd1542ebc14ec5dd0837283